requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
ansicon==1.89.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
blessed==1.20.0
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
contexttimer==0.3.3
datadispatch==1.0.0
Deprecated==1.2.14
discord==2.3.2
discord-py-interactions==4.4.1
discord-ui==5.1.6
discord.py==2.3.2
docopt==0.6.2
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
humanfriendly==10.0
humanize==4.8.0
idna==3.4
jinxed==1.2.0
lavalink==4.0.6
lyricsgenius==3.0.1
multidict==6.0.4
mutagen==1.47.0
numpy==1.26.0
packaging==23.1
pafy==0.5.5
pandas==2.1.1
pipreqs==0.4.13
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.1.4
requests==2.31.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.5
spotipy==2.23.0
srsly==2.4.8
tasks==2.8.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.5
wavelink==1.3.5
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
youtube-dl==2021.12.17
youtube-search==2.1.2
yt-dlp==2023.9.24
